Extract hotbackup related functions into separate file.
