+++ arangod/Cluster/ClusterHotBackup.h
@@ -1,0 +1,126 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2014-2024 ArangoDB GmbH, Cologne, Germany
+/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
+///
+/// Licensed under the Business Source License 1.1 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     https://github.com/arangodb/arangodb/blob/devel/LICENSE
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is ArangoDB GmbH, Cologne, Germany
+///
+/// @author Max Neunhoeffer
+/// @author Kaveh Vahedipour
+////////////////////////////////////////////////////////////////////////////////
+
+#pragma once
+
+#include "Cluster/ClusterTypes.h"
+
+#include <velocypack/Slice.h>
+
+#include <map>
+#include <string>
+#include <vector>
+
+namespace arangodb {
+class ClusterFeature;
+class Result;
+namespace velocypack {
+class Builder;
+}
+
+//////////////////////////////////////////////////////////////////////////////
+/// @brief create hotbackup on a coordinator
+//////////////////////////////////////////////////////////////////////////////
+
+enum HotBackupMode { CONSISTENT, DIRTY };
+
+/**
+ * @brief Create hot backup on coordinators
+ * @param mode    Backup mode: consistent, dirty
+ * @param timeout Wait for this attempt and bail out if not met
+ */
+arangodb::Result hotBackupCoordinator(ClusterFeature&, VPackSlice const payload,
+                                      VPackBuilder& report);
+
+/**
+ * @brief Restore specific hot backup on coordinators
+ * @param mode    Backup mode: consistent, dirty
+ * @param timeout Wait for this attempt and bail out if not met
+ */
+arangodb::Result hotRestoreCoordinator(ClusterFeature&,
+                                       VPackSlice const payload,
+                                       VPackBuilder& report);
+
+/**
+ * @brief List all
+ * @param mode    Backup mode: consistent, dirty
+ * @param timeout Wait for this attempt and bail out if not met
+ */
+arangodb::Result listHotBackupsOnCoordinator(ClusterFeature&,
+                                             VPackSlice const payload,
+                                             VPackBuilder& report);
+
+/**
+ * @brief Delete specific hot backup
+ * @param backupId  BackupId to delete
+ */
+arangodb::Result deleteHotBackupsOnCoordinator(ClusterFeature&,
+                                               VPackSlice const payload,
+                                               VPackBuilder& report);
+
+#ifdef USE_ENTERPRISE
+/**
+ * @brief Trigger upload of specific hot backup
+ * @param backupId  BackupId to delete
+ */
+arangodb::Result uploadBackupsOnCoordinator(ClusterFeature&,
+                                            VPackSlice const payload,
+                                            VPackBuilder& report);
+
+/**
+ * @brief Trigger download of specific hot backup
+ * @param backupId  BackupId to delete
+ */
+arangodb::Result downloadBackupsOnCoordinator(ClusterFeature&,
+                                              VPackSlice const payload,
+                                              VPackBuilder& report);
+#endif
+
+/**
+ * @brief match backup servers
+ * @param  planDump   Dump of plan from backup
+ * @param  dbServers  This cluster's db servers
+ * @param  match      Matched db servers
+ * @return            Operation's success
+ */
+arangodb::Result matchBackupServers(VPackSlice const planDump,
+                                    std::vector<ServerID> const& dbServers,
+                                    std::map<std::string, std::string>& match);
+
+arangodb::Result matchBackupServersSlice(VPackSlice const planServers,
+                                         std::vector<ServerID> const& dbServers,
+                                         std::map<ServerID, ServerID>& match);
+
+/**
+ * @brief apply database server matches to plan
+ * @param  plan     Plan from hot backup
+ * @param  matches  Match backup's server ids to new server ids
+ * @param  newPlan  Resulting new plan
+ * @return          Operation's result
+ */
+arangodb::Result applyDBServerMatchesToPlan(
+    VPackSlice const plan, std::map<ServerID, ServerID> const& matches,
+    VPackBuilder& newPlan);
+
+}  // namespace arangodb
