+++ arangod/Cluster/ClusterMethods.h
@@ -195,90 +195,6 @@ Result compactOnAllDBServers(ClusterFeature&, bool changeLevel,
 Result compactOnAllDBServers(ClusterFeature&, std::string const& dbname,
                              std::string const& collname);
 
-//////////////////////////////////////////////////////////////////////////////
-/// @brief create hotbackup on a coordinator
-//////////////////////////////////////////////////////////////////////////////
-
-enum HotBackupMode { CONSISTENT, DIRTY };
-
-/**
- * @brief Create hot backup on coordinators
- * @param mode    Backup mode: consistent, dirty
- * @param timeout Wait for this attempt and bail out if not met
- */
-arangodb::Result hotBackupCoordinator(ClusterFeature&, VPackSlice const payload,
-                                      VPackBuilder& report);
-
-/**
- * @brief Restore specific hot backup on coordinators
- * @param mode    Backup mode: consistent, dirty
- * @param timeout Wait for this attempt and bail out if not met
- */
-arangodb::Result hotRestoreCoordinator(ClusterFeature&,
-                                       VPackSlice const payload,
-                                       VPackBuilder& report);
-
-/**
- * @brief List all
- * @param mode    Backup mode: consistent, dirty
- * @param timeout Wait for this attempt and bail out if not met
- */
-arangodb::Result listHotBackupsOnCoordinator(ClusterFeature&,
-                                             VPackSlice const payload,
-                                             VPackBuilder& report);
-
-/**
- * @brief Delete specific hot backup
- * @param backupId  BackupId to delete
- */
-arangodb::Result deleteHotBackupsOnCoordinator(ClusterFeature&,
-                                               VPackSlice const payload,
-                                               VPackBuilder& report);
-
-#ifdef USE_ENTERPRISE
-/**
- * @brief Trigger upload of specific hot backup
- * @param backupId  BackupId to delete
- */
-arangodb::Result uploadBackupsOnCoordinator(ClusterFeature&,
-                                            VPackSlice const payload,
-                                            VPackBuilder& report);
-
-/**
- * @brief Trigger download of specific hot backup
- * @param backupId  BackupId to delete
- */
-arangodb::Result downloadBackupsOnCoordinator(ClusterFeature&,
-                                              VPackSlice const payload,
-                                              VPackBuilder& report);
-#endif
-
-/**
- * @brief match backup servers
- * @param  planDump   Dump of plan from backup
- * @param  dbServers  This cluster's db servers
- * @param  match      Matched db servers
- * @return            Operation's success
- */
-arangodb::Result matchBackupServers(VPackSlice const planDump,
-                                    std::vector<ServerID> const& dbServers,
-                                    std::map<std::string, std::string>& match);
-
-arangodb::Result matchBackupServersSlice(VPackSlice const planServers,
-                                         std::vector<ServerID> const& dbServers,
-                                         std::map<ServerID, ServerID>& match);
-
-/**
- * @brief apply database server matches to plan
- * @param  plan     Plan from hot backup
- * @param  matches  Match backup's server ids to new server ids
- * @param  newPlan  Resulting new plan
- * @return          Operation's result
- */
-arangodb::Result applyDBServerMatchesToPlan(
-    VPackSlice const plan, std::map<ServerID, ServerID> const& matches,
-    VPackBuilder& newPlan);
-
 /// @brief get the engine stats from all DB servers
 arangodb::Result getEngineStatsFromDBServers(ClusterFeature&,
                                              VPackBuilder& report);
